Completed FeedbackServices and part of EventServices

+++ src/Application/OrganizingEventsService.Application.Models/Dto/Feedback/FeedbackDto.cs
@@ -4,7 +4,7 @@ public class FeedbackDto
 {
     public Guid Id { get; set; }
 
-    public uint Rating { get; set; }
+    public decimal Rating { get; set; }
 
     public string? Text { get; set; }
